requirements.txt
asgiref==3.2.4
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
requests==2.5.0
six==1.14.0
sqlparse==0.3.1
swapi==0.1.3
typed-ast==1.4.1
ujson==1.33
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
